Refactor StartScreen into smaller widgets and services
